requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
werkzeug>=2.3.3
beautifulsoup4==4.9.3
asgiref==3.4.1
Pillow==10.2.0
python-docx==0.8.11
numpy==1.21.6
pandas==1.3.5
openai>=1.0.0
